setup permission in genre

+++ app/Http/Controllers/Portals/GenreController.php
@@ -38,13 +38,17 @@ class GenreController extends Controller
                 })
                 ->addColumn('action', function ($row) {
                     $btn = '<a href="'. route('portal.genres.show', $row->guid) .'" data-id="'.$row->guid.'" class="btn btn-success btn-sm mb-1 mr-1"><i class="far fa-eye"></i></a>';
-                    $btn .= '<a href="'. route('portal.genres.edit', $row->guid) .'" data-id="'.$row->guid.'" class="btn btn-primary btn-sm mb-1"><i class="far fa-edit"></i></a>';
-                    $btn .= ' <button onclick="deleteData('. "'$row->guid'" .')" data-id="'.$row->guid.'" class="btn btn-danger btn-sm mb-1"><i class="far fa-trash-alt"></i></button>';
-                    $btn .= '<form id="deleteForm'. $row->guid .'" action="'. route('portal.genres.destroy', $row->guid) .'" method="POST" style="display: none">
-                    <input type="hidden" name="_token" value="'. csrf_token() .'">
-                    <input type="hidden" name="_method" value="DELETE">
-                    @method("DELETE")
-                    </form>';
+                    if (auth()->user()->hasRole('SuperAdmin') || auth()->user()->hasPermissionTo('editgenre')) {
+                        $btn .= '<a href="'. route('portal.genres.edit', $row->guid) .'" data-id="'.$row->guid.'" class="btn btn-primary btn-sm mb-1"><i class="far fa-edit"></i></a>';
+                    }
+                    if (auth()->user()->hasRole('SuperAdmin') || auth()->user()->hasPermissionTo('deletegenre')) {
+                        $btn .= ' <button onclick="deleteData('. "'$row->guid'" .')" data-id="'.$row->guid.'" class="btn btn-danger btn-sm mb-1"><i class="far fa-trash-alt"></i></button>';
+                        $btn .= '<form id="deleteForm'. $row->guid .'" action="'. route('portal.genres.destroy', $row->guid) .'" method="POST" style="display: none">
+                        <input type="hidden" name="_token" value="'. csrf_token() .'">
+                        <input type="hidden" name="_method" value="DELETE">
+                        @method("DELETE")
+                        </form>';
+                    }
 
                     return $btn;
                 })
